requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
django-appconf==1.0.5
django-imagekit==5.0.0
djangorestframework==3.14.0
pilkit==3.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3

# deployment module(s)
gunicorn

# Reformat and Test
black
flake8